+++ src/main/java/duke/parse/command/AddCommand.java
@@ -1,0 +1,18 @@
+package duke.parse.command;
+
+import duke.Duke;
+import duke.task.Task;
+
+public class AddCommand implements Command {
+    private Task task;
+
+    public AddCommand(Task task) {
+        this.task = task;
+    }
+
+    @Override
+    public boolean execute(Duke bot) {
+        bot.addTaskToList(task);
+        return true;
+    }
+}
