+++ src/main/java/duke/parse/command/DeleteCommand.java
@@ -1,0 +1,17 @@
+package duke.parse.command;
+
+import duke.Duke;
+
+public class DeleteCommand implements Command {
+    private int taskIndex;
+
+    public DeleteCommand(int taskIndex) {
+        this.taskIndex = taskIndex;
+    }
+
+    @Override
+    public boolean execute(Duke bot) {
+        bot.deleteTask(this.taskIndex);
+        return true;
+    }
+}
