+++ src/main/java/duke/parse/command/EmptyCommand.java
@@ -1,0 +1,12 @@
+package duke.parse.command;
+
+import duke.Duke;
+
+public class EmptyCommand implements Command {
+    public EmptyCommand() {}
+
+    @Override
+    public boolean execute(Duke bot) {
+        return true;
+    }
+}
