+++ src/main/java/duke/parse/command/ExitCommand.java
@@ -1,0 +1,14 @@
+package duke.parse.command;
+
+import duke.Duke;
+
+public class ExitCommand implements Command {
+    public ExitCommand() {
+    }
+
+    @Override
+    public boolean execute(Duke bot) {
+        bot.exit();
+        return false;
+    }
+}
