+++ src/main/java/duke/parse/command/SaveCommand.java
@@ -1,0 +1,13 @@
+package duke.parse.command;
+
+import duke.Duke;
+
+public class SaveCommand implements Command {
+    public SaveCommand() {}
+
+    @Override
+    public boolean execute(Duke bot) {
+        bot.saveData();
+        return true;
+    }
+}
