+++ src/main/java/duke/task/Event.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.parse.DateTimeManager;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
