+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 import java.time.LocalDate;
 
 public abstract class Task {
